Changes based on comments to PR 12

+++ dataactvalidator/validation_handlers/validator.py
@@ -1,4 +1,5 @@
 import re
+from decimal import *
 from dataactcore.utils.responseException import ResponseException
 from dataactcore.utils.statusCode import StatusCode
 from dataactvalidator.validation_handlers.validationError import ValidationError
@@ -126,7 +127,7 @@ class Validator(object):
                 return False
         if(datatype == "DECIMAL") :
             try:
-                float(data)
+                Decimal(data)
                 return True
             except:
                 return False
@@ -157,7 +158,7 @@ class Validator(object):
         if(datatype =="INT") :
             return int(float(data))
         if(datatype =="DECIMAL") :
-            return float(data)
+            return Decimal(data)
         if(datatype == "STRING" or datatype =="BOOLEAN") :
             return data
         if(datatype == "LONG"):
